+++ bot.js
@@ -68,8 +68,23 @@ client.getRandom = (len) => {
 client.convertColor = (rrggbb) => {
     var newColor = rrggbb.substr(1);
     return parseInt(newColor, 16);
-    };
+  };
 
+//to truncate strings to stay within the 250 character embed limit
+client.trunc = (str, len, end) => {
+  if(len == null) {
+    len == 250;
+  }
+  if(end == null) {
+    end == "...";
+  }
+  if(str.length > len) {
+    return(str.substring(0,len - end.length) + end);
+  }
+  else {
+    return str;
+  }
+}
 /* -------- end functions -------- */