+++ CHANGELOG.md
@@ -10,6 +10,12 @@ For any questions or to get in contact, message me at
 LOG
 ---
 
+[v 0.0.7]
+Cleanup
+
+- Added a function to truncate strings
+- Used the above function to limit the length of image descriptions as embed titles to stay within Discord's 250 character limit
+
 [v 0.0.6]
 Error handling